chore: Add a GitHub action that runs unit tests with -race to CI build (#4774) (#4775)
* chore: Add a GitHub action that runs unit tests with -race to CI build (#4774)


* chore: Add a GitHub action that runs unit tests with -race to CI build (#4774)
