+++ util/session/sessionmanager_test.go
@@ -233,26 +233,6 @@ func TestCacheValueGetters(t *testing.T) {
 
 }
 
-func TestRandomPasswordVerificationDelay(t *testing.T) {
-	var sleptFor time.Duration
-	settingsMgr := settings.NewSettingsManager(context.Background(), getKubeClient("password", true), "argocd")
-	mgr := newSessionManager(settingsMgr, NewInMemoryUserStateStorage())
-	mgr.verificationDelayNoiseEnabled = true
-	mgr.sleep = func(d time.Duration) {
-		sleptFor = d
-	}
-	for i := 0; i < 10; i++ {
-		sleptFor = 0
-		start := time.Now()
-		if !assert.NoError(t, mgr.VerifyUsernamePassword("admin", "password")) {
-			return
-		}
-		totalDuration := time.Since(start) + sleptFor
-		assert.GreaterOrEqual(t, totalDuration.Nanoseconds(), verificationDelayNoiseMin.Nanoseconds())
-		assert.LessOrEqual(t, totalDuration.Nanoseconds(), verificationDelayNoiseMax.Nanoseconds())
-	}
-}
-
 func TestLoginRateLimiter(t *testing.T) {
 	settingsMgr := settings.NewSettingsManager(context.Background(), getKubeClient("password", true), "argocd")
 	storage := NewInMemoryUserStateStorage()
